refactor scripts used for openapi (#30309)
